luv-setup: Tiny wrapper to make configuration easier
We currently have far too many configuration steps that are required
to build a luv image. And if you miss any of those steps, the resulting
failures are not at all helpful in diagnosing what you did wrong.

We specify a luv-specific TEMPLATECONF directory where luv-specific layers
and distro are given out of the box.

This patch addresses:
https://github.com/01org/luv-yocto/issues/31

Originally-by: Matt Fleming <matt@codeblueprint.co.uk>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>