luv-test-parser: Sanitize plain text for HTML report
When we incorporate the output of the unit tests of each test suite, we
need to ensure that such text renders correctly in the generated HTML
report. Specifically, certain characters need to be escaped or modified
because they have special meaning in HTML. For instance, the characters
"<" and ">" are used to define markup tags. Nothing prevents test suites
from using these characters in their output. Thus, we need to replace
these characters with the appropriate HTML entities.

A special function, html_sanitize_text(), is defined with the goal of
grouping together the text substitutions that need to be performed in
text intended to be included in the HTML report. This function can be
easily extended in the future if needed.

While here, relocate the substitution of the new line character ("\n")
into the new function html_sanitize_text().

This fixes https://github.com/01org/luv-yocto/issues/118.

Cc: Naresh Bhat <naresh.bhat@linaro.org>
Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>