luv-test: Send messages to console and debug file
Currently, we only use the plymouth graphical manager to display
certain messages to the user. Hence, someone with access to only
serial console or netconsole will not know what is going on. We
need to write these messages to a console.

Also, when the luv.debug command line parameter is passed, we
write all the plymouth messages to console as well as the luv-debug
file. We also need to export the necessary variables.

Signed-off-by: Megha Dey <megha.dey@linux.intel.com>