telemetrics: Run pstore-clean *only* when luv.telemetrics is enabled and in kexec kernel
pstore-clean service, copies crash contents out of pstore directory and
cleans up the directory; so as to free up the space. LUV writes crash
log to efi variables in normal kernel (if it crashes) and collects crash
report *only* in kexec kernel and sends it to telemetrics server only
when luv.telemetrics is enabled. Hence, add these conditions to
pstore-clean service file.

Cc: Naresh Bhat <naresh.bhat@linaro.org>
Cc: Megha Dey <megha.dey@linux.intel.com>
Cc: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Sai Praneeth Prakhya <sai.praneeth.prakhya@intel.com>